requirements.txt
pyusb
pynput
